Refactor quantity and batch information details
